brand-new-to-do/requirements.txt
Django==2.2.1
djangorestframework==3.10.1
psycopg2-binary==2.8.2
pytz==2019.1
sqlparse==0.3.0
social-auth-app-django==3.1.0